Add failing test for error checking on 64 bits int
Currently some error checking in buf.go is done by converting an uint32 to
int. For 32 bit int this works correctly, however when int is 64 bits,
0xffffffff != -1 and this error checking fails, leading to a slice out of
bounds panic.

Some editorialization has been done to use a VALUES literal as opposed
to temp table/inserts in the test.

+++ buf.go
@@ -8,7 +8,7 @@ import (
 type readBuf []byte
 
 func (b *readBuf) int32() (n int) {
-	n = int(binary.BigEndian.Uint32(*b))
+	n = int(int32(binary.BigEndian.Uint32(*b)))
 	*b = (*b)[4:]
 	return
 }
